+++ tasks/gulp/__tests__/after-build-webjar.test.js
@@ -137,25 +137,4 @@ describe('webjar/', () => {
 
     expect(govukFrontendVersion).toEqual(expectedGovUkVersion);
   });
-
-  it('should have the correct accessible-autocompleted groupId', async () => {
-    const govukFrontendGroupId = await getHmrcPomField('/project/dependencies/dependency[2]/groupId');
-
-    expect(govukFrontendGroupId).toEqual('org.webjars.npm');
-  });
-
-  it('should have the correct accessible-autocomplete artifactId', async () => {
-    const accessibleAutocomplete = await getHmrcPomField('/project/dependencies/dependency[2]/artifactId');
-
-    expect(accessibleAutocomplete).toEqual('accessible-autocomplete');
-  });
-
-  it('should have the correct accessible-autocomplete version, ignoring any ~ or ^ qualifiers', async () => {
-    const expectedAutocompleteVersion = packagePackageJson.dependencies['accessible-autocomplete']
-      .replace(/[^~]/, '');
-
-    const autocompleteVersion = await getHmrcPomField('/project/dependencies/dependency[2]/version');
-
-    expect(autocompleteVersion).toEqual(expectedAutocompleteVersion);
-  });
 });